refactor: review fixes — DRY, ergonomics, dead code removal
- Remove dead SharedData<T> (superseded by Arc<Event>)
- Update LazyConversation comment (Arc dispatch gives parse-once for free)
- Add Event::as_message() helper for clean destructuring
- Add MessageContext::from_event() to eliminate construction boilerplate
- Extract duplicated TestEventCollector to test_utils
- Simplify double-match patterns in e2e tests with as_message()
